Рефакторит улиты
